+++ src/main/java/fr/openent/lystore/export/equipmentRapp/AnnexeDelibTab.java
@@ -72,7 +72,7 @@ public class AnnexeDelibTab extends TabHelper {
         int lineToInsert = 5;
         JsonObject idPassed = new JsonObject();
         String key = "", oldkey = "";
-        Float oldTotal = 0.f;
+        Double oldTotal = 0.d;
         for (int i = 0; i < datas.size(); i++) {
             JsonObject action = datas.getJsonObject(i);
             key = action.getString("program") + " - " + action.getString("code");
@@ -87,10 +87,10 @@ public class AnnexeDelibTab extends TabHelper {
                 oldkey = "";
             }
             if (!oldkey.equals(key)) {
-                oldTotal = 0.f;
+                oldTotal = 0.d;
             }
             oldkey = key;
-            oldTotal += Float.parseFloat(action.getString("total"));
+            oldTotal += Double.parseDouble(action.getString("total"));
 
             excel.insertCellTabDouble(columnToInsert + 4, lineToInsert, oldTotal);
